refactor: split napi transform helpers
