+++ crates/ox_content_napi/src/tabs.rs
@@ -13,7 +13,15 @@
 //! itself stays on the JS side. The exact output is pinned by the
 //! `embed-transform` characterization tests in `@ox-content/vite-plugin`.
 
+mod render;
+mod scan;
+#[cfg(test)]
+mod tests;
+
 use crate::html_scan::find_ci;
+
+use render::render_tabs;
+use scan::{attribute_value, find_matching_close, find_tag, scan_start_tag};
 
 struct Tab {
     label: String,
@@ -111,291 +119,4 @@ fn parse_tabs(inner: &str) -> Vec<Tab> {
         cursor = close_start + "</tab>".len();
     }
     tabs
-}
-
-fn render_tabs(tabs: &[Tab], group: u32) -> String {
-    use std::fmt::Write;
-
-    // The group number is interpolated many times below; format it once and
-    // reuse the string instead of allocating per use. The per-tab index is
-    // written directly into `out` via `write!`, avoiding a temporary `String`
-    // per number.
-    let group_str = group.to_string();
-
-    let mut out = String::new();
-    out.push_str("<div class=\"ox-tabs-container\"><div class=\"ox-tabs\" data-group=\"");
-    out.push_str(&group_str);
-    out.push_str("\"><div class=\"ox-tabs-header\">");
-    for (index, tab) in tabs.iter().enumerate() {
-        out.push_str("<input type=\"radio\" name=\"ox-tabs-");
-        out.push_str(&group_str);
-        out.push_str("\" id=\"ox-tab-");
-        let _ = write!(out, "{group_str}-{index}");
-        out.push('"');
-        if index == 0 {
-            out.push_str(" checked");
-        }
-        out.push_str("><label for=\"ox-tab-");
-        let _ = write!(out, "{group_str}-{index}");
-        out.push_str("\">");
-        out.push_str(&escape_text(&tab.label));
-        out.push_str("</label>");
-    }
-    out.push_str("</div>");
-    for (index, tab) in tabs.iter().enumerate() {
-        out.push_str("<div class=\"ox-tab-panel\" data-tab=\"");
-        let _ = write!(out, "{index}");
-        out.push_str("\">");
-        out.push_str(&tab.content);
-        out.push_str("</div>");
-    }
-    out.push_str("</div><noscript><div class=\"ox-tabs-fallback\">");
-    for (index, tab) in tabs.iter().enumerate() {
-        out.push_str("<details");
-        if index == 0 {
-            out.push_str(" open");
-        }
-        out.push_str("><summary>");
-        out.push_str(&escape_text(&tab.label));
-        out.push_str("</summary><div class=\"ox-tabs-fallback-content\">");
-        out.push_str(&tab.content);
-        out.push_str("</div></details>");
-    }
-    out.push_str("</div></noscript></div>");
-    out
-}
-
-/// A scanned start tag.
-struct StartTag {
-    /// Byte index just past the element name (start of the attribute region).
-    name_end: usize,
-    /// Byte index of the attribute region end (before `/` of `/>` or before `>`).
-    inner_end: usize,
-    /// Byte index just past the closing `>`.
-    end: usize,
-    self_closing: bool,
-}
-
-/// Scan the start tag beginning at `pos` (which must point at `<`), respecting
-/// quoted attribute values so a `>` inside a value doesn't end the tag early.
-fn scan_start_tag(html: &str, pos: usize) -> Option<StartTag> {
-    let bytes = html.as_bytes();
-    if bytes.get(pos) != Some(&b'<') {
-        return None;
-    }
-    let mut i = pos + 1;
-    while i < bytes.len() && !bytes[i].is_ascii_whitespace() && bytes[i] != b'>' && bytes[i] != b'/'
-    {
-        i += 1;
-    }
-    let name_end = i;
-    let mut quote: Option<u8> = None;
-    let mut tag_end = None;
-    while i < bytes.len() {
-        let b = bytes[i];
-        match quote {
-            Some(q) => {
-                if b == q {
-                    quote = None;
-                }
-            }
-            None => {
-                if b == b'"' || b == b'\'' {
-                    quote = Some(b);
-                } else if b == b'>' {
-                    tag_end = Some(i);
-                    break;
-                }
-            }
-        }
-        i += 1;
-    }
-    let tag_end = tag_end?;
-    let self_closing = tag_end > pos && bytes[tag_end - 1] == b'/';
-    let inner_end = if self_closing { tag_end - 1 } else { tag_end };
-    Some(StartTag { name_end, inner_end, end: tag_end + 1, self_closing })
-}
-
-/// Find the next start tag at or after `from` with a proper element boundary.
-/// `open` is the `<name` literal (e.g. `"<tabs"`); the boundary check means
-/// `<tab` never matches `<tabs`/`<table` and `<tabs` never matches `<tabset>`.
-fn find_tag(html: &str, from: usize, open: &str) -> Option<usize> {
-    let bytes = html.as_bytes();
-    let mut search = from;
-    loop {
-        let at = find_ci(html, search, open)?;
-        let after = at + open.len();
-        let boundary = bytes.get(after).copied();
-        if matches!(boundary, Some(b) if b == b'>' || b == b'/' || b.is_ascii_whitespace()) {
-            return Some(at);
-        }
-        search = after;
-    }
-}
-
-/// Given the position just past a `<name …>` start tag, find the byte offset of
-/// its matching `close` literal, accounting for nested `open` (`<name`) tags.
-fn find_matching_close(html: &str, from: usize, open: &str, close: &str) -> Option<usize> {
-    let mut depth = 1usize;
-    let mut search = from;
-    loop {
-        let next_open = find_tag(html, search, open);
-        let next_close = find_ci(html, search, close);
-        match (next_open, next_close) {
-            (Some(open_at), Some(close_at)) if open_at < close_at => {
-                depth += 1;
-                // Advance past this nested open tag's `>`.
-                search = match scan_start_tag(html, open_at) {
-                    Some(tag) => tag.end,
-                    None => open_at + 1,
-                };
-            }
-            (_, Some(close_at)) => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(close_at);
-                }
-                search = close_at + close.len();
-            }
-            (_, None) => return None,
-        }
-    }
-}
-
-/// Read the value of `name` (case-insensitive) from a start tag's attribute
-/// region. Supports quoted and unquoted values; missing value yields `""`.
-fn attribute_value(attrs: &str, name: &str) -> Option<String> {
-    // The transform only needs one requested attribute from a start-tag slice.
-    // A narrow byte scan is faster and simpler than constructing a generic
-    // attribute map for every tab element.
-    let bytes = attrs.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        while i < bytes.len() && bytes[i].is_ascii_whitespace() {
-            i += 1;
-        }
-        if i >= bytes.len() || bytes[i] == b'/' {
-            break;
-        }
-        let name_start = i;
-        while i < bytes.len()
-            && !bytes[i].is_ascii_whitespace()
-            && bytes[i] != b'='
-            && bytes[i] != b'/'
-        {
-            i += 1;
-        }
-        let attr_name = &attrs[name_start..i];
-        while i < bytes.len() && bytes[i].is_ascii_whitespace() {
-            i += 1;
-        }
-        let value = if i < bytes.len() && bytes[i] == b'=' {
-            i += 1;
-            while i < bytes.len() && bytes[i].is_ascii_whitespace() {
-                i += 1;
-            }
-            if i >= bytes.len() {
-                String::new()
-            } else if bytes[i] == b'"' || bytes[i] == b'\'' {
-                let q = bytes[i];
-                i += 1;
-                let vs = i;
-                while i < bytes.len() && bytes[i] != q {
-                    i += 1;
-                }
-                let v = attrs[vs..i].to_string();
-                if i < bytes.len() {
-                    i += 1;
-                }
-                v
-            } else {
-                let vs = i;
-                while i < bytes.len() && !bytes[i].is_ascii_whitespace() && bytes[i] != b'/' {
-                    i += 1;
-                }
-                attrs[vs..i].to_string()
-            }
-        } else {
-            String::new()
-        };
-        if attr_name.eq_ignore_ascii_case(name) {
-            return Some(value);
-        }
-    }
-    None
-}
-
-/// Escape text content the way `hast-util-to-html` does for the characters that
-/// can occur in tab labels.
-fn escape_text(value: &str) -> String {
-    let mut out = String::with_capacity(value.len());
-    for ch in value.chars() {
-        match ch {
-            '&' => out.push_str("&#x26;"),
-            '<' => out.push_str("&#x3C;"),
-            '>' => out.push_str("&#x3E;"),
-            _ => out.push(ch),
-        }
-    }
-    out
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn expands_two_tabs_matching_characterization() {
-        let result = transform_tabs(
-            r#"<tabs><tab title="A">alpha</tab><tab title="B">beta</tab></tabs>"#,
-            0,
-        );
-        assert_eq!(result.group_count, 1);
-        assert_eq!(
-            result.html,
-            r#"<div class="ox-tabs-container"><div class="ox-tabs" data-group="0"><div class="ox-tabs-header"><input type="radio" name="ox-tabs-0" id="ox-tab-0-0" checked><label for="ox-tab-0-0">Tab 1</label><input type="radio" name="ox-tabs-0" id="ox-tab-0-1"><label for="ox-tab-0-1">Tab 2</label></div><div class="ox-tab-panel" data-tab="0">alpha</div><div class="ox-tab-panel" data-tab="1">beta</div></div><noscript><div class="ox-tabs-fallback"><details open><summary>Tab 1</summary><div class="ox-tabs-fallback-content">alpha</div></details><details><summary>Tab 2</summary><div class="ox-tabs-fallback-content">beta</div></details></div></noscript></div>"#
-        );
-    }
-
-    #[test]
-    fn uses_label_attribute_when_present() {
-        let result = transform_tabs(r#"<tabs><tab label="First">x</tab></tabs>"#, 0);
-        assert!(result.html.contains("<label for=\"ox-tab-0-0\">First</label>"));
-        assert!(result.html.contains("<summary>First</summary>"));
-    }
-
-    #[test]
-    fn numbers_groups_from_start_and_counts() {
-        let result =
-            transform_tabs(r"<tabs><tab>a</tab></tabs> middle <tabs><tab>b</tab></tabs>", 5);
-        assert_eq!(result.group_count, 2);
-        assert!(result.html.contains(r#"data-group="5""#));
-        assert!(result.html.contains(r#"data-group="6""#));
-        assert!(result.html.contains(" middle "));
-    }
-
-    #[test]
-    fn leaves_empty_tabs_untouched_and_uncounted() {
-        let html = r"<tabs>   </tabs>";
-        let result = transform_tabs(html, 0);
-        assert_eq!(result.group_count, 0);
-        assert_eq!(result.html, html);
-    }
-
-    #[test]
-    fn passes_through_without_tabs_marker() {
-        let html = r"<p>No tabs here, just a <table><tr><td>cell</td></tr></table>.</p>";
-        let result = transform_tabs(html, 0);
-        assert_eq!(result.group_count, 0);
-        assert_eq!(result.html, html);
-    }
-
-    #[test]
-    fn preserves_rich_inner_content() {
-        let result =
-            transform_tabs(r#"<tabs><tab label="Code"><pre><code>x</code></pre></tab></tabs>"#, 0);
-        assert!(result
-            .html
-            .contains(r#"<div class="ox-tab-panel" data-tab="0"><pre><code>x</code></pre></div>"#));
-    }
 }
